% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_objects.R
\name{CounterOptions}
\alias{CounterOptions}
\title{CounterOptions Object}
\usage{
CounterOptions(metric = NULL, field = NULL)
}
\arguments{
\item{metric}{The metric to update}

\item{field}{The field value to attribute}
}
\value{
CounterOptions object
}
\description{
CounterOptions Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Options for counters
}
