% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_objects.R
\name{ListReposResponse}
\alias{ListReposResponse}
\title{ListReposResponse Object}
\usage{
ListReposResponse(repos = NULL)
}
\arguments{
\item{repos}{The listed repos}
}
\value{
ListReposResponse object
}
\description{
ListReposResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for ListRepos.
}
