% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_objects.R
\name{MirrorConfig}
\alias{MirrorConfig}
\title{MirrorConfig Object}
\usage{
MirrorConfig(url = NULL, webhookId = NULL, deployKeyId = NULL)
}
\arguments{
\item{url}{URL of the main repository at the other hosting service}

\item{webhookId}{ID of the webhook listening to updates to trigger mirroring}

\item{deployKeyId}{ID of the SSH deploy key at the other hosting service}
}
\value{
MirrorConfig object
}
\description{
MirrorConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Configuration to automatically mirror a repository from anotherhosting service, for example GitHub or BitBucket.
}
