% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_functions.R
\name{projects.repos.create}
\alias{projects.repos.create}
\title{Creates a repo in the given project with the given name..If the named repository already exists, `CreateRepo` returns`ALREADY_EXISTS`.}
\usage{
projects.repos.create(Repo, parent)
}
\arguments{
\item{Repo}{The \link{Repo} object to pass to this method}

\item{parent}{The project in which to create the repo}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/eap/cloud-repositories/cloud-sourcerepo-api}{Google Documentation}

Other Repo functions: \code{\link{Repo}}
}
