% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_functions.R
\name{projects.repos.delete}
\alias{projects.repos.delete}
\title{Deletes a repo.}
\usage{
projects.repos.delete(name)
}
\arguments{
\item{name}{The name of the repo to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/eap/cloud-repositories/cloud-sourcerepo-api}{Google Documentation}
}
