% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_functions.R
\name{projects.repos.setIamPolicy}
\alias{projects.repos.setIamPolicy}
\title{Sets the access control policy on the specified resource. Replaces anyexisting policy.}
\usage{
projects.repos.setIamPolicy(SetIamPolicyRequest, resource)
}
\arguments{
\item{SetIamPolicyRequest}{The \link{SetIamPolicyRequest} object to pass to this method}

\item{resource}{REQUIRED: The resource for which the policy is being specified}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/eap/cloud-repositories/cloud-sourcerepo-api}{Google Documentation}

Other SetIamPolicyRequest functions: \code{\link{SetIamPolicyRequest}}
}
