% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_objects.R
\name{Repo}
\alias{Repo}
\title{Repo Object}
\usage{
Repo(mirrorConfig = NULL, url = NULL, size = NULL, name = NULL)
}
\arguments{
\item{mirrorConfig}{How this repository mirrors a repository managed by another service}

\item{url}{URL to clone the repository from Google Cloud Source Repositories}

\item{size}{The size in bytes of the repo}

\item{name}{Resource name of the repository, of the form}
}
\value{
Repo object
}
\description{
Repo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A repository (or repo) is a Git repository storing versioned source content.
}
\seealso{
Other Repo functions: \code{\link{projects.repos.create}}
}
