% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_objects.R
\name{SetIamPolicyRequest}
\alias{SetIamPolicyRequest}
\title{Cloud Source Repositories API Objects 
Access source code repositories hosted by Google.}
\usage{
SetIamPolicyRequest(policy = NULL, updateMask = NULL)
}
\arguments{
\item{policy}{REQUIRED: The complete policy to be applied to the `resource`}

\item{updateMask}{OPTIONAL: A FieldMask specifying which fields of the policy to modify}
}
\value{
SetIamPolicyRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:14:46
filename: /Users/mark/dev/R/autoGoogleAPI/googlesourcerepov1.auto/R/sourcerepo_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
SetIamPolicyRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for `SetIamPolicy` method.
}
\seealso{
Other SetIamPolicyRequest functions: \code{\link{projects.repos.setIamPolicy}}
}
