% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcerepo_functions.R
\docType{package}
\name{sourcerepo_googleAuthR}
\alias{sourcerepo_googleAuthR}
\alias{sourcerepo_googleAuthR-package}
\title{Cloud Source Repositories API
Access source code repositories hosted by Google.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:14:45
filename: /Users/mark/dev/R/autoGoogleAPI/googlesourcerepov1.auto/R/sourcerepo_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
